# docs/openapi.yml
...
openapi: 3.0.1
info:
  title: Dompet API
  description: API for managing banking operations
  version: 2.0.1
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /transfer:
    get:
      tags:
        - Transfer
      summary: Get all transfers
      description: Retrieve a list of all transfers
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transfer'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Transfer
      summary: Create a new transfer
      description: Create a new transfer between accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transfer'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /transferRecipient:
    get:
      tags:
        - TransferRecipient
      summary: Get all transfer recipients
      description: Retrieve a list of all transfer recipients
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransferRecipient'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - TransferRecipient
      summary: Create a new transfer recipient
      description: Create a new transfer recipient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRecipient'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferRecipient'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /transaction:
    get:
      tags:
        - Transaction
      summary: Get all transactions
      description: Retrieve a list of all transactions
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Transaction
      summary: Create a new transaction
      description: Create a new transaction for an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /account/{accountId}/transfers/{transferId}:
    get:
      tags:
        - Account-Transfer
      summary: Find a transaction by ID for an account
      description: Retrieve a transaction associated with an account
      parameters:
        - name: accountId
          in: path
          required: true
          description: The ID of the account
          schema:
            type: string
        - name: transferId
          in: path
          required: true
          description: The ID of the transfer
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: string
                example: 'Category updated successfully'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /account/{accountId}/transactions/{transactionId}:
    get:
      tags:
        - Account-Transaction
      summary: Find a transaction by ID for an account
      description: Retrieve a transaction associated with an account
      parameters:
        - name: accountId
          in: path
          required: true
          description: The ID of the account
          schema:
            type: string
        - name: transactionId
          in: path
          required: true
          description: The ID of the transaction
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: string
                example: 'Category updated successfully'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /transferRecipient/{id}:
    get:
      tags:
        - TransferRecipient
      summary: Get a transfer recipient by ID
      description: Retrieve a transfer recipient by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the transfer recipient
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferRecipient'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - TransferRecipient
      summary: Delete a transfer recipient
      description: Delete a transfer recipient by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the transfer recipient
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /transfer/{id}:
    get:
      tags:
        - Transfer
      summary: Get a transfer by ID
      description: Retrieve a transfer by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the transfer
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Transfer
      summary: Delete a transfer
      description: Delete a transfer by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the transfer
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /transfer/{id}/transferRecipient:
    get:
      tags:
        - Transfer
      summary: Get transfer recipients for a transfer
      description: Retrieve a list of transfer recipients associated with a transfer
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the transfer
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransferRecipient'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /transaction/{id}:
    get:
      tags:
        - Transaction
      summary: Get a transaction by ID
      description: Retrieve a transaction by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the transaction
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Transaction
      summary: Delete a transaction
      description: Delete a transaction by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the transaction
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 401
            message: Unauthorized
    NotFoundError:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 404
            message: Not Found
    InternalServerError:
      description: An unexpected error occurred on the server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 500
            message: Internal Server Error
    BadRequestError:
      description: The request was malformed or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            message: Bad Request
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          description: The HTTP status code
        message:
          type: string
          description: A human-readable error message
    Transfer:
      type: object
      description: Represents a transfer between accounts
      required:
        - scheduledEffectiveDate
        - senderAccountId
        - uniqueReference
      properties:
        transferId:
          type: string
          maxLength: 50
          description: The unique identifier of the transfer
        uniqueReference:
          type: string
          maxLength: 20
          description: A unique reference for the transfer
        scheduledEffectiveDate:
          type: string
          format: date
          description: The scheduled effective date of the transfer
        statusId:
          type: integer
          format: int32
          description: The status ID of the transfer
        description:
          type: string
          description: A description of the transfer
        effectiveDate:
          type: string
          format: date-time
          description: The effective date of the transfer
        recordDate:
          type: string
          format: date-time
          description: The date the transfer was recorded
        senderAccountId:
          type: string
          description: The ID of the account sending the transfer
        isInternal:
          type: boolean
          description: Indicates if the transfer is internal to the bank
      example:
        transferId: 'TR12345'
        uniqueReference: 'REF12345'
        scheduledEffectiveDate: '2023-05-01'
        statusId: 1
        description: 'Monthly salary transfer'
        effectiveDate: '2023-05-01T12:00:00Z'
        recordDate: '2023-04-30T14:30:00Z'
        senderAccountId: 'AC12345'
        isInternal: true
    TransferRecipient:
      type: object
      description: Represents a recipient of a transfer
      required:
        - recipientAccountId
        - transferId
      properties:
        transferRecipientId:
          type: integer
          format: int64
          description: The unique identifier of the transfer recipient
        transferId:
          type: string
          description: The ID of the associated transfer
        recipientAccountId:
          type: string
          description: The ID of the recipient account
        amount:
          type: number
          description: The amount transferred to the recipient
      example:
        transferRecipientId: 1
        transferId: 'TR12345'
        recipientAccountId: 'AC67890'
        amount: 100000
    Transaction:
      type: object
      description: Represents a transaction associated with an account
      required:
        - accountId
        - amount
        - categoryId
        - effectiveDate
      properties:
        transactionId:
          type: string
          maxLength: 50
          description: The unique identifier of the transaction
        accountId:
          type: string
          description: The ID of the associated account
        categoryId:
          type: string
          description: The ID of the category for the transaction
        effectiveDate:
          type: string
          format: date-time
          description: The effective date of the transaction
        recordDate:
          type: string
          format: date-time
          description: The date the transaction was recorded
        amount:
          type: number
          description: The amount of the transaction
        description:
          type: string
          description: A description of the transaction
      example:
        transactionId: 'TX12345'
        accountId: 'AC12345'
        categoryId: 'CAT1'
        effectiveDate: '2023-04-01T10:00:00Z'
        recordDate: '2023-04-01T12:30:00Z'
        amount: 500
        description: 'Grocery shopping'
...